Apply the M5 adversarial review's findings
Four of the six review dimensions reported; each finding was checked against
the code before it was acted on. Three were defects that produced wrong
Markdown and that every fixture and every unit check had missed:

- A line end inside a w:t reached the output as a line end, so
  <w:t>Total&#10;# 5</w:t> came out as a paragraph followed by a heading and a
  pretty-printed w:t came out as an indented code block. WordprocessingML
  spells a break w:br; a newline inside a w:t is interior whitespace and now
  folds to one space.
- A GFM delimiter row could attach to the line above it inside one paragraph,
  so a|b with a hard break and then -|- rendered as a table. The line-start
  pass now escapes the head of anything shaped like a delimiter row.
- A thematic break whose hyphens were not one contiguous run was not escaped;
  CommonMark counts three or more hyphens with any spacing between them.

Also from the review: w:caps now uppercases its run's text, w:webHidden hides
a run the way w:vanish does, w:dir/w:bdo/w:ruby's w:rubyBase are descended
into rather than skipped, DocFindStyle no longer copies indeterminate bytes
into its cache, the dead mc:ProcessContent element branch is gone (MCE spells
it an attribute), a raw unsigned long long cast left an alias-only TU, five
public accessors gained the @return d1 requires, and the wrappers fixture's
w:tbl and dangling r:id were made valid WordprocessingML.

Fixed in the documentation: the unit-check count, the tests/unit file count,
a claim that mc:ProcessContent is transparent when no such branch exists, and
the breakdown of run_golden.py's 47 checks. The accessor documentation
convention d1 leaves open is now written down rather than merely practised.

D12 is raised and open: GitHub renders $...$ as math, which
docs/CONVERSION_REFERENCE.md 4.1 predates, so whether to escape $ is a ruling
rather than a fix.

Verified on Linux against the shim build, all under AddressSanitizer and
UndefinedBehaviorSanitizer with leak detection on and no diagnostic: the unit
suite 809 checks, tests/run_golden.py 47, tests/run_container.py 99;
4,000 documents through an independent CommonMark renderer, 3,150 style
tables against an independent implementation of ISO/IEC 29500-1 17.7.3,
15,000 mutated archives and 7,500 generated documents, all clean.
clang-format --style=file is a no-op on all 41 C++ files.

Not run here, because MSVC is the only supported compiler and this is Linux:
  msbuild DOCXtoMD.sln /m /p:Configuration=Release /p:Platform=x64
  msbuild DOCXtoMD.sln /m /p:Configuration=Debug   /p:Platform=x64
  python tests\make_fixtures.py
  python tests\run_container.py
  python tests\run_golden.py
  tests\x64\Release\DOCXtoMD.Tests.exe

+++ src/Convert.cpp
@@ -305,7 +305,7 @@ cEXIT_CODE ConvertFile(cCLI_OPTIONSptr options, cwchptr inputPath) {
       if(verdict == EXIT_ALL_CONVERTED && !options->quiet) {
          char note[64];
 
-         snprintf(note, sizeof(note), "wrote %llu bytes", (unsigned long long)MdByteCount(&emitter));
+         snprintf(note, sizeof(note), "wrote %llu bytes", MdByteCount(&emitter));
          DiagNoteText(note, outputPath);
       }
    }
